added routes for register, login, logout, refresh old tokens

+++ src/utils/cloudinary.js
@@ -1,37 +1,32 @@
-import {v2 as cloudinary} from "cloudinary";
-import fs from "fs";
+import {v2 as cloudinary} from "cloudinary"
+import fs from "fs"
 
-cloudinary.config({
-    cloud_name: process.env.CLOUDINARY_CLOUD_NAME,
-    api_key: process.env.CLOUDINARY_API_KEY,
-    api_secret: process.env.CLOUDINARY_API_SECRET
-});
+
+
 
 const uploadOnCloudinary = async (localFilePath) => {
+    console.log("check: ",localFilePath)
     try {
-        if(!localFilePath) return null;
-
-        //upload local file on cloudinary
+        if (!localFilePath) return null
+        //upload the file on cloudinary
         const response = await cloudinary.uploader.upload(localFilePath, {
-            //to be filled later
             resource_type: "auto"
         })
-
-        //file has been uploaded successfully
-        console.log("file uploaded successfully on cloudinary: ", response.url);
-
-        //unlink file
-
-
+        // file has been uploaded successfull
+        //console.log("file is uploaded on cloudinary ", response.url);
+        // fs.unlinkSync(localFilePath)
         return response;
 
     } catch (error) {
-        fs.unlinkSync(localFilePath); // remove the locally saved temporary file as upload operation failed;
-
-
-        return null;
+    console.error("❌ Cloudinary Upload Error:", error.message);  // <-- this will show the reason
+    if (fs.existsSync(localFilePath)) {
+        fs.unlinkSync(localFilePath);
     }
+    return null;
+}
+
 }
 
 
-export {uploadOnCloudinary};
+
+export {uploadOnCloudinary, cloudinary }
